+++ test/raxx/router_test.exs
@@ -51,6 +51,15 @@ defmodule Raxx.RouterTest do
     end
   end
 
+  defmodule InvalidReturn do
+    use Raxx.Server
+
+    @impl Raxx.Server
+    def handle_request(_request, _state) do
+      :foo
+    end
+  end
+
   defmodule MyRouter do
     use Raxx.Server
 
@@ -59,31 +68,33 @@ defmodule Raxx.RouterTest do
       {%{method: :GET, path: ["users"]}, UsersPage},
       {%{method: :GET, path: ["users", _id]}, UserPage},
       {%{method: :POST, path: ["users"]}, CreateUser},
+      {%{method: :GET, path: ["invalid"]}, InvalidReturn},
+      {%{method: :POST, path: ["invalid"]}, InvalidReturn},
       {_, NotFoundPage}
     ]
   end
 
   test "will route to homepage" do
     request = Raxx.request(:GET, "/")
-    response = MyRouter.handle_head(request, :state)
+    {[response], _state} = MyRouter.handle_head(request, :state)
     assert "Home page" == response.body
   end
 
   test "will route to fixed segment" do
     request = Raxx.request(:GET, "/users")
-    response = MyRouter.handle_head(request, :state)
+    {[response], _state} = MyRouter.handle_head(request, :state)
     assert "Users page" == response.body
   end
 
   test "will route to variable segment path" do
     request = Raxx.request(:GET, "/users/34")
-    response = MyRouter.handle_head(request, :state)
+    {[response], _state} = MyRouter.handle_head(request, :state)
     assert "User page 34" == response.body
   end
 
   test "will route on method" do
     request = Raxx.request(:POST, "/users")
-    response = MyRouter.handle_head(request, :state)
+    {[response], _state} = MyRouter.handle_head(request, :state)
     assert "User created " == response.body
   end
 
@@ -94,13 +105,13 @@ defmodule Raxx.RouterTest do
 
     {[], state} = MyRouter.handle_head(request, :state)
     {[], state} = MyRouter.handle_data("Bob", state)
-    response = MyRouter.handle_tail([], state)
+    {[response], _state} = MyRouter.handle_tail([], state)
     assert "User created Bob" == response.body
   end
 
   test "will route on catch all" do
     request = Raxx.request(:GET, "/random")
-    response = MyRouter.handle_head(request, :state)
+    {[response], _state} = MyRouter.handle_head(request, :state)
     assert "Not found" == response.body
   end
 
@@ -110,5 +121,26 @@ defmodule Raxx.RouterTest do
     metadata = Logger.metadata()
     assert "Raxx.RouterTest.HomePage" = Keyword.get(metadata, :"raxx.action")
     assert "%{method: :GET, path: []}" = Keyword.get(metadata, :"raxx.route")
+  end
+
+  test "will raise return error if fails to route simple request" do
+    request = Raxx.request(:GET, "/invalid")
+
+    assert_raise ReturnError, fn ->
+      MyRouter.handle_head(request, :state)
+    end
+  end
+
+  test "will raise return error if fails to route streamed request" do
+    request =
+      Raxx.request(:POST, "/invalid")
+      |> Raxx.set_body(true)
+
+    {[], state} = MyRouter.handle_head(request, :state)
+    {[], state} = MyRouter.handle_data("Bob", state)
+
+    assert_raise ReturnError, fn ->
+      MyRouter.handle_tail([], state)
+    end
   end
 end