+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## master
+
+### Fixed
+
+- Router will return informative return error like any server module.
+
 ## [0.15.2](https://github.com/CrowdHailer/raxx/tree/0.15.2) - 2018-04-28
 
 ### Added